+++ projects/myrmidon/cadmus-part-general-pg/src/lib/token-text-part-feature/token-text-part-feature.component.spec.ts
@@ -2,8 +2,6 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
-import { NgeMonacoModule } from '@cisstech/nge/monaco';
-
 import { TokenTextPartFeatureComponent } from './token-text-part-feature.component';
 import { CurrentItemBarComponent } from '@myrmidon/cadmus-item-editor';
 import {
@@ -21,7 +19,6 @@ describe('TokenTextPartFeatureComponent', () => {
         CommonModule,
         FormsModule,
         ReactiveFormsModule,
-        NgeMonacoModule,
         TokenTextPartFeatureComponent,
       ],
       providers: [