+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # History
 
-- 2026-06-12: updated Angular and packages.
+- 2026-06-12:
+  - updated Angular and packages.
+  - ⚠️ migrated from Monaco editor wrapper `@cisstech/nge` to a new wrapper `@jean-merelis/ngx-monaco-editor` + `marked` for Markdown rendering, so to get rid of `@cisstech/nge` altogether.
 - 2026-06-08: ⚠️ upgraded to Angular 22 bumping all libraries major version.
 - 2026-05-20:
   - added ID display to event type.